+++ CMakeLists.txt
@@ -325,9 +325,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL "Linux")
 endif()
 
 if(CMAKE_SYSTEM_NAME STREQUAL "NetBSD")
-  list(APPEND uv_sources
-       src/unix/netbsd.c
-       src/unix/random-sysctl-bsd.c)
+  list(APPEND uv_sources src/unix/netbsd.c)
 endif()
 
 if(CMAKE_SYSTEM_NAME STREQUAL "OpenBSD")
