--- src/unix/random-sysctl-bsd.c
@@ -1,51 +1,0 @@
-/* Copyright libuv contributors. All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to
- * deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- */
-
-#include "uv.h"
-#include "internal.h"
-
-#include <errno.h>
-#include <string.h>
-
-#include <sys/sysctl.h>
-#include <unistd.h>
-
-int uv__random_sysctl(void* buf, size_t len) {
-  static int name[] = {CTL_KERN, KERN_ARND};
-  size_t count, req;
-  unsigned char *p = buf;
-
-  while (len) {
-    req = len < 32 ? len : 32;
-    count = req;
-
-    if (sysctl(name, ARRAY_SIZE(name), p, &count, NULL, 0) == -1)
-      return UV__ERR(errno);
-
-    if (count != req)
-      return UV_EIO;  /* Can't happen. */
-
-    p += count;
-    len -= count;
-  }
-
-  return 0;
-}
